Move SplitConfigTest to test_kastManip
