+++ pyk/src/tests/test_kastManip.py
@@ -12,6 +12,7 @@ from pyk.kastManip import (
     push_down_rewrites,
     remove_generated_cells,
     simplify_bool,
+    splitConfigFrom,
     substitute,
 )
 from pyk.prelude import Bool, Sorts, intToken, mlAnd, mlEqualsTrue, mlTop
@@ -260,3 +261,18 @@ class BuildClaimtest(TestCase):
                 target_cterm = CTerm(target)
                 kclaim, _ = build_claim('claim', init_cterm, target_cterm)
                 self.assertEqual(kclaim, claim)
+
+
+class SplitConfigTest(TestCase):
+    def test_splitConfigFrom(self):
+        k_cell = KSequence([KApply('foo'), KApply('bar')])
+        term = KApply('<k>', [k_cell])
+        config, subst = splitConfigFrom(term)
+        self.assertEqual(config, KApply('<k>', [KVariable('K_CELL')]))
+        self.assertEqual(subst, {'K_CELL': k_cell})
+
+        map_item_cell = KApply('<mapItem>', [KApply('foo')])
+        map_cell = KApply('<mapCell>', [KApply('map_join', [map_item_cell, map_item_cell])])
+        config, subst = splitConfigFrom(map_cell)
+        self.assertEqual(config, KApply('<mapCell>', [KVariable('MAPCELL_CELL')]))
+        self.assertEqual(subst, {'MAPCELL_CELL': KApply('map_join', [map_item_cell, map_item_cell])})
